+++ src/client/Microsoft.Identity.Client/WsTrust/WsTrustWebRequestManager.cs
@@ -21,6 +21,10 @@ namespace Microsoft.Identity.Client.WsTrust
     internal class WsTrustWebRequestManager : IWsTrustWebRequestManager
     {
         private readonly IHttpManager _httpManager;
+        private readonly LinearRetryPolicy _linearRetryPolicy = new LinearRetryPolicy(
+            LinearRetryPolicy.DefaultStsRetryDelayMs,
+            LinearRetryPolicy.DefaultStsMaxRetries,
+            HttpRetryConditions.Sts);
 
         public WsTrustWebRequestManager(IHttpManager httpManager)
         {
@@ -44,15 +48,17 @@ namespace Microsoft.Identity.Client.WsTrust
             var uri = new UriBuilder(federationMetadataUrl);
 
             HttpResponse httpResponse = await _httpManager.SendRequestAsync(
-                    uri.Uri,
-                    msalIdParams,
-                    body: null,
-                    method: HttpMethod.Get,
-                    logger: requestContext.Logger,
-                    doNotThrow: false,
-                    mtlsCertificate: null,
-                    validateServerCertificate: null, cancellationToken: requestContext.UserCancellationToken)
-                .ConfigureAwait(false);
+                uri.Uri,
+                msalIdParams,
+                body: null,
+                method: HttpMethod.Get,
+                logger: requestContext.Logger,
+                doNotThrow: false,
+                mtlsCertificate: null,
+                validateServerCertificate: null,
+                cancellationToken: requestContext.UserCancellationToken,
+                retryPolicy: _linearRetryPolicy)
+            .ConfigureAwait(false);
 
             if (httpResponse.StatusCode != System.Net.HttpStatusCode.OK)
             {
@@ -94,21 +100,23 @@ namespace Microsoft.Identity.Client.WsTrust
                 { "SOAPAction", (wsTrustEndpoint.Version == WsTrustVersion.WsTrust2005) ? XmlNamespace.Issue2005.ToString() : XmlNamespace.Issue.ToString() }
             };
 
-            // CodeQL [SM00417] False Positive: wsTrustRequest is a body parameter for HttpRequest that follows WsTrust protocol
             var body = new StringContent(
+                // CodeQL [SM00417] False Positive: wsTrustRequest is a body parameter for HttpRequest that follows WsTrust protocol
                 wsTrustRequest,
                 Encoding.UTF8, "application/soap+xml");
 
             HttpResponse resp = await _httpManager.SendRequestAsync(
-                    wsTrustEndpoint.Uri,
-                    headers,
-                    body: body,
-                    method: HttpMethod.Post,
-                    logger: requestContext.Logger,
-                    doNotThrow: true,
-                    mtlsCertificate: null,
-                    validateServerCertificate: null, cancellationToken: requestContext.UserCancellationToken)
-                .ConfigureAwait(false);
+                wsTrustEndpoint.Uri,
+                headers,
+                body: body,
+                method: HttpMethod.Post,
+                logger: requestContext.Logger,
+                doNotThrow: true,
+                mtlsCertificate: null,
+                validateServerCertificate: null,
+                cancellationToken: requestContext.UserCancellationToken,
+                retryPolicy: _linearRetryPolicy)
+            .ConfigureAwait(false);
 
             if (resp.StatusCode != System.Net.HttpStatusCode.OK)
             {
@@ -175,15 +183,17 @@ namespace Microsoft.Identity.Client.WsTrust
             var uri = new UriBuilder(userRealmUriPrefix + userName + "?api-version=1.0").Uri;
 
             var httpResponse = await _httpManager.SendRequestAsync(
-                   uri,
-                   msalIdParams,
-                   body: null,
-                   method: HttpMethod.Get,
-                   logger: requestContext.Logger,
-                   doNotThrow: false,
-                   mtlsCertificate: null,
-                   validateServerCertificate: null, cancellationToken: requestContext.UserCancellationToken)
-                .ConfigureAwait(false);
+                uri,
+                msalIdParams,
+                body: null,
+                method: HttpMethod.Get,
+                logger: requestContext.Logger,
+                doNotThrow: false,
+                mtlsCertificate: null,
+                validateServerCertificate: null,
+                cancellationToken: requestContext.UserCancellationToken,
+                retryPolicy: _linearRetryPolicy)
+            .ConfigureAwait(false);
 
             if (httpResponse.StatusCode == System.Net.HttpStatusCode.OK)
             {